+++ Specs/Core/EllipsoidRhumbLineSpec.js
@@ -1,11 +1,11 @@
 defineSuite([
-        'Core/EllipsoidRhumb',
+        'Core/EllipsoidRhumbLine',
         'Core/Cartographic',
         'Core/Ellipsoid',
         'Core/EllipsoidGeodesic',
         'Core/Math'
     ], function(
-        EllipsoidRhumb,
+        EllipsoidRhumbLine,
         Cartographic,
         Ellipsoid,
         EllipsoidGeodesic,
@@ -14,28 +14,28 @@ defineSuite([
 
     it('throws without start', function() {
         expect(function() {
-            var rhumb = new EllipsoidRhumb();
+            var rhumb = new EllipsoidRhumbLine();
             return rhumb.interpolateUsingSurfaceDistance(0);
         }).toThrowDeveloperError();
     });
 
     it('throws without end', function() {
         expect(function() {
-            var rhumb = new EllipsoidRhumb(new Cartographic(Math.PI, Math.PI));
+            var rhumb = new EllipsoidRhumbLine(new Cartographic(Math.PI, Math.PI));
             return rhumb.interpolateUsingSurfaceDistance(0);
         }).toThrowDeveloperError();
     });
 
     it('throws without unique position', function() {
         expect(function() {
-            var rhumb = new EllipsoidRhumb(new Cartographic(Math.PI, Math.PI), new Cartographic(0, Math.PI));
+            var rhumb = new EllipsoidRhumbLine(new Cartographic(Math.PI, Math.PI), new Cartographic(0, Math.PI));
             return rhumb.interpolateUsingSurfaceDistance(0);
         }).toThrowDeveloperError();
     });
 
     it('setEndPoints throws without start', function() {
         expect(function() {
-            var rhumb = new EllipsoidRhumb();
+            var rhumb = new EllipsoidRhumbLine();
             rhumb.setEndPoints();
         }).toThrowDeveloperError();
     });
@@ -43,7 +43,7 @@ defineSuite([
     it('setEndPoints throws without end', function() {
         expect(function() {
             var start = new Cartographic(CesiumMath.PI_OVER_TWO, 0);
-            var rhumb = new EllipsoidRhumb();
+            var rhumb = new EllipsoidRhumbLine();
             rhumb.setEndPoints(start);
             return rhumb.interpolateUsingSurfaceDistance(0);
         }).toThrowDeveloperError();
@@ -55,13 +55,13 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var end = new Cartographic(thirtyDegrees, thirtyDegrees);
 
-        var rhumb = EllipsoidRhumb.fromStartAndEnd(start, end);
+        var rhumb = EllipsoidRhumbLine.fromStartAndEnd(start, end);
         expect(start).toEqual(rhumb.start);
         expect(end).toEqual(rhumb.end);
     });
 
     it('can create using fromStartAndEnd function with result', function() {
-        var scratch = new EllipsoidRhumb();
+        var scratch = new EllipsoidRhumbLine();
 
         var ellipsoid = new Ellipsoid(6, 6, 3);
         var fifteenDegrees = Math.PI / 12;
@@ -69,7 +69,7 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var end = new Cartographic(thirtyDegrees, thirtyDegrees);
 
-        var rhumb = EllipsoidRhumb.fromStartAndEnd(start, end, ellipsoid, scratch);
+        var rhumb = EllipsoidRhumbLine.fromStartAndEnd(start, end, ellipsoid, scratch);
         expect(rhumb).toBe(scratch);
         expect(rhumb.ellipsoid).toBe(ellipsoid);
         expect(start).toEqual(rhumb.start);
@@ -83,14 +83,14 @@ defineSuite([
         var heading = fifteenDegrees;
         var distance = 6 * fifteenDegrees;
 
-        var rhumb = EllipsoidRhumb.fromStartHeadingDistance(start, heading, distance, ellipsoid);
+        var rhumb = EllipsoidRhumbLine.fromStartHeadingDistance(start, heading, distance, ellipsoid);
         expect(start).toEqual(rhumb.start);
         expect(distance).toEqualEpsilon(rhumb.surfaceDistance, CesiumMath.EPSILON12);
         expect(heading).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
     });
 
     it('can create using fromStartHeadingDistance function with result', function() {
-        var scratch = new EllipsoidRhumb();
+        var scratch = new EllipsoidRhumbLine();
 
         var ellipsoid = new Ellipsoid(6, 6, 3);
         var fifteenDegrees = Math.PI / 12;
@@ -98,7 +98,7 @@ defineSuite([
         var heading = fifteenDegrees;
         var distance = 6 * fifteenDegrees;
 
-        var rhumb = EllipsoidRhumb.fromStartHeadingDistance(start, heading, distance, ellipsoid, scratch);
+        var rhumb = EllipsoidRhumbLine.fromStartHeadingDistance(start, heading, distance, ellipsoid, scratch);
         expect(rhumb).toBe(scratch);
         expect(rhumb.ellipsoid).toBe(ellipsoid);
         expect(start).toEqual(rhumb.start);
@@ -108,14 +108,14 @@ defineSuite([
 
     it('getSurfaceDistance throws if start or end never defined', function() {
         expect(function() {
-            var rhumb = new EllipsoidRhumb();
+            var rhumb = new EllipsoidRhumbLine();
             return rhumb.surfaceDistance;
         }).toThrowDeveloperError();
     });
 
     it('getHeading throws if start or end never defined', function() {
         expect(function() {
-            var rhumb = new EllipsoidRhumb();
+            var rhumb = new EllipsoidRhumbLine();
             return rhumb.heading;
         }).toThrowDeveloperError();
     });
@@ -126,7 +126,7 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var end = new Cartographic(thirtyDegrees, thirtyDegrees);
 
-        var rhumb = new EllipsoidRhumb(start, end);
+        var rhumb = new EllipsoidRhumbLine(start, end);
         expect(start).toEqual(rhumb.start);
         expect(end).toEqual(rhumb.end);
     });
@@ -134,7 +134,7 @@ defineSuite([
     it('sets end points', function() {
         var start = new Cartographic(CesiumMath.PI_OVER_TWO, 0);
         var end = new Cartographic(CesiumMath.PI_OVER_TWO, CesiumMath.PI_OVER_TWO);
-        var rhumb = new EllipsoidRhumb();
+        var rhumb = new EllipsoidRhumbLine();
         rhumb.setEndPoints(start, end);
         expect(start).toEqual(rhumb.start);
         expect(end).toEqual(rhumb.end);
@@ -145,7 +145,7 @@ defineSuite([
         var start = new Cartographic(CesiumMath.PI_OVER_TWO, 0);
         var end = new Cartographic(Math.PI, 0);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
         expect(CesiumMath.PI_OVER_TWO).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
     });
 
@@ -154,7 +154,7 @@ defineSuite([
         var start = new Cartographic(0, 1.2);
         var end = new Cartographic(Math.PI, 1.5);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
 
         expect(0.0).not.toEqual(rhumb.heading);
     });
@@ -164,7 +164,7 @@ defineSuite([
         var start = new Cartographic(1.3, CesiumMath.PI_OVER_TWO);
         var end = new Cartographic(0.0, CesiumMath.PI / 2.4);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
 
         expect(0.0).not.toEqual(rhumb.heading);
     });
@@ -174,13 +174,13 @@ defineSuite([
         var start = new Cartographic(0.0, 0.3);
         var end = new Cartographic(CesiumMath.PI_OVER_TWO, 0.3);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
 
         expect(CesiumMath.PI_OVER_TWO).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
 
         start = new Cartographic(3 * CesiumMath.PI_OVER_TWO, 0.3);
         end = new Cartographic(CesiumMath.PI, 0.3);
-        var rhumb2 = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb2 = new EllipsoidRhumbLine(start, end, ellipsoid);
         expect(-CesiumMath.PI_OVER_TWO).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
     });
 
@@ -189,10 +189,10 @@ defineSuite([
         var start = new Cartographic(0.0, 1.2);
         var end = new Cartographic(0.0, 1.5);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
         expect(0.0).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
 
-        var rhumb2 = new EllipsoidRhumb(end, start, ellipsoid);
+        var rhumb2 = new EllipsoidRhumbLine(end, start, ellipsoid);
         expect(CesiumMath.PI).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
     });
 
@@ -201,7 +201,7 @@ defineSuite([
         var start = new Cartographic(CesiumMath.PI_OVER_TWO, 0.0);
         var end = new Cartographic(CesiumMath.PI_OVER_TWO, 0.0);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
         expect(0).toEqualEpsilon(rhumb.surfaceDistance, CesiumMath.EPSILON12);
         expect(0).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
     });
@@ -211,7 +211,7 @@ defineSuite([
         var start = new Cartographic(-CesiumMath.PI_OVER_FOUR, 0.0);
         var end = new Cartographic(CesiumMath.PI_OVER_FOUR, 0.0);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
         expect(CesiumMath.PI_OVER_TWO * 6).toEqualEpsilon(rhumb.surfaceDistance, CesiumMath.EPSILON12);
     });
 
@@ -222,7 +222,7 @@ defineSuite([
         var fortyfiveDegrees = Math.PI / 4;
         var end = new Cartographic(CesiumMath.PI_OVER_TWO, fortyfiveDegrees);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
         var thirtyDegrees = Math.PI / 6;
         expect(thirtyDegrees * 6).toEqualEpsilon(rhumb.surfaceDistance, CesiumMath.EPSILON12);
     });
@@ -234,10 +234,10 @@ defineSuite([
         var fortyFiveWest = new Cartographic(-CesiumMath.PI_OVER_FOUR, 0.0);
         var fortyFiveEast = new Cartographic(CesiumMath.PI_OVER_FOUR, 0.0);
 
-        var westEastRhumb = new EllipsoidRhumb(fortyFiveWest, fortyFiveEast, ellipsoid);
-        var southNorthRhumb = new EllipsoidRhumb(fortyFiveSouth, fortyFiveNorth, ellipsoid);
-        var eastWestRhumb = new EllipsoidRhumb(fortyFiveEast, fortyFiveWest, ellipsoid);
-        var northSouthRhumb = new EllipsoidRhumb(fortyFiveNorth, fortyFiveSouth, ellipsoid);
+        var westEastRhumb = new EllipsoidRhumbLine(fortyFiveWest, fortyFiveEast, ellipsoid);
+        var southNorthRhumb = new EllipsoidRhumbLine(fortyFiveSouth, fortyFiveNorth, ellipsoid);
+        var eastWestRhumb = new EllipsoidRhumbLine(fortyFiveEast, fortyFiveWest, ellipsoid);
+        var northSouthRhumb = new EllipsoidRhumbLine(fortyFiveNorth, fortyFiveSouth, ellipsoid);
         expect(CesiumMath.PI_OVER_TWO * 6).toEqualEpsilon(westEastRhumb.surfaceDistance, CesiumMath.EPSILON12);
         expect(CesiumMath.PI_OVER_TWO * 6).toEqualEpsilon(southNorthRhumb.surfaceDistance, CesiumMath.EPSILON12);
         expect(westEastRhumb.surfaceDistance).toEqualEpsilon(southNorthRhumb.surfaceDistance, CesiumMath.EPSILON12);
@@ -253,7 +253,7 @@ defineSuite([
         var start = new Cartographic(0, -fortyfiveDegrees);
         var end = new Cartographic(CesiumMath.PI_OVER_TWO, -fortyfiveDegrees);
 
-        var rhumb = new EllipsoidRhumb(start, end, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(start, end, ellipsoid);
         var distance = Math.cos(fortyfiveDegrees) * CesiumMath.PI_OVER_TWO * 6;
         expect(distance).toEqualEpsilon(rhumb.surfaceDistance, CesiumMath.EPSILON12);
     });
@@ -265,8 +265,8 @@ defineSuite([
         var initial = new Cartographic(fifteenDegrees, fifteenDegrees);
         var distance = radius * fifteenDegrees;
 
-        var rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, fifteenDegrees, distance, ellipsoid);
-        var rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        var rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, fifteenDegrees, distance, ellipsoid);
+        var rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
 
         expect(fifteenDegrees).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(distance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
@@ -278,8 +278,8 @@ defineSuite([
         var initial = new Cartographic(fifteenDegrees, fifteenDegrees);
         var distance = ellipsoid.maximumRadius * fifteenDegrees;
 
-        var rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, fifteenDegrees, distance, ellipsoid);
-        var rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        var rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, fifteenDegrees, distance, ellipsoid);
+        var rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
 
         expect(fifteenDegrees).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(distance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
@@ -299,28 +299,28 @@ defineSuite([
         var ninetyPointOneDegrees = 90.1 * oneDegree;
         var ninetyPointZeroTwoDegrees = 90.02 * oneDegree;
 
-        var rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, eightyNineDegrees, distance, ellipsoid);
-        var rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        var rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, eightyNineDegrees, distance, ellipsoid);
+        var rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
 
-        rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, eightyNinePointNineDegrees, distance, ellipsoid);
-        rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, eightyNinePointNineDegrees, distance, ellipsoid);
+        rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
 
-        rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, ninetyDegrees, distance, ellipsoid);
-        rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, ninetyDegrees, distance, ellipsoid);
+        rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
 
-        rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, ninetyPointOneDegrees, distance, ellipsoid);
-        rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, ninetyPointOneDegrees, distance, ellipsoid);
+        rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
 
-        rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, ninetyPointZeroTwoDegrees, distance, ellipsoid);
-        rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, ninetyPointZeroTwoDegrees, distance, ellipsoid);
+        rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
     });
@@ -338,28 +338,28 @@ defineSuite([
         var ninetyPointOneDegrees = 90.1 * oneDegree;
         var ninetyPointZeroTwoDegrees = 90.02 * oneDegree;
 
-        var rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, eightyNineDegrees, distance, ellipsoid);
-        var rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        var rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, eightyNineDegrees, distance, ellipsoid);
+        var rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON3);
 
-        rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, eightyNinePointNineDegrees, distance, ellipsoid);
-        rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, eightyNinePointNineDegrees, distance, ellipsoid);
+        rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON3);
 
-        rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, ninetyDegrees, distance, ellipsoid);
-        rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, ninetyDegrees, distance, ellipsoid);
+        rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON3);
 
-        rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, ninetyPointOneDegrees, distance, ellipsoid);
-        rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, ninetyPointOneDegrees, distance, ellipsoid);
+        rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON3);
 
-        rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, ninetyPointZeroTwoDegrees, distance, ellipsoid);
-        rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, ninetyPointZeroTwoDegrees, distance, ellipsoid);
+        rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON3);
     });
@@ -371,8 +371,8 @@ defineSuite([
         var final = new Cartographic(fifteenDegrees, 0.0);
         var distance = ellipsoid.maximumRadius * 2 * fifteenDegrees;
 
-        var rhumb1 = new EllipsoidRhumb(initial, final, ellipsoid);
-        var rhumb2 = EllipsoidRhumb.fromStartHeadingDistance(initial, CesiumMath.PI_OVER_TWO, distance, ellipsoid);
+        var rhumb1 = new EllipsoidRhumbLine(initial, final, ellipsoid);
+        var rhumb2 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, CesiumMath.PI_OVER_TWO, distance, ellipsoid);
 
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
@@ -386,16 +386,16 @@ defineSuite([
 
         var distance = ellipsoid.maximumRadius * 2 * fifteenDegrees;
 
-        var rhumb1 = new EllipsoidRhumb(initial, final, ellipsoid);
-        var rhumb2 = new EllipsoidRhumb.fromStartHeadingDistance(initial, 3.0 * CesiumMath.PI_OVER_TWO, distance, ellipsoid);
+        var rhumb1 = new EllipsoidRhumbLine(initial, final, ellipsoid);
+        var rhumb2 = new EllipsoidRhumbLine.fromStartHeadingDistance(initial, 3.0 * CesiumMath.PI_OVER_TWO, distance, ellipsoid);
 
         expect(-CesiumMath.PI_OVER_TWO).toEqualEpsilon(rhumb1.heading, CesiumMath.EPSILON12);
         expect(CesiumMath.PI / 6 * ellipsoid.maximumRadius).toEqualEpsilon(rhumb1.surfaceDistance, CesiumMath.EPSILON6);
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
 
-        var rhumb3 = new EllipsoidRhumb(final, initial, ellipsoid);
-        var rhumb4 = new EllipsoidRhumb.fromStartHeadingDistance(final, CesiumMath.PI_OVER_TWO, distance, ellipsoid);
+        var rhumb3 = new EllipsoidRhumbLine(final, initial, ellipsoid);
+        var rhumb4 = new EllipsoidRhumbLine.fromStartHeadingDistance(final, CesiumMath.PI_OVER_TWO, distance, ellipsoid);
         expect(CesiumMath.PI_OVER_TWO).toEqualEpsilon(rhumb3.heading, CesiumMath.EPSILON12);
         expect(CesiumMath.PI / 6 * ellipsoid.maximumRadius).toEqualEpsilon(rhumb3.surfaceDistance, CesiumMath.EPSILON6);
         expect(rhumb3.heading).toEqualEpsilon(rhumb4.heading, CesiumMath.EPSILON12);
@@ -410,7 +410,7 @@ defineSuite([
         var final = new Cartographic(fifteenDegrees, oneDegree);
 
         //A rhumb line with heading = 0 should be almost the same as a geodesic
-        var rhumb = new EllipsoidRhumb(initial, final, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(initial, final, ellipsoid);
         var geodesic = new EllipsoidGeodesic(initial, final, ellipsoid);
         expect(0.0).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
         expect(geodesic.startHeading).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
@@ -422,7 +422,7 @@ defineSuite([
         var final = new Cartographic(CesiumMath.PI - 1, 0.0);
 
         //A rhumb line on the equator should be the same as a geodesic
-        var rhumb = new EllipsoidRhumb(initial, final, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(initial, final, ellipsoid);
         var geodesic = new EllipsoidGeodesic(initial, final, ellipsoid);
         expect(CesiumMath.PI_OVER_TWO).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
         expect(geodesic.startHeading).toEqualEpsilon(rhumb.heading, CesiumMath.EPSILON12);
@@ -439,8 +439,8 @@ defineSuite([
 
         var initial = new Cartographic(0.0, eightyDegrees);
 
-        var rhumb1 = EllipsoidRhumb.fromStartHeadingDistance(initial, eightyDegrees, distance, ellipsoid);
-        var rhumb2 = new EllipsoidRhumb(initial, rhumb1.end, ellipsoid);
+        var rhumb1 = EllipsoidRhumbLine.fromStartHeadingDistance(initial, eightyDegrees, distance, ellipsoid);
+        var rhumb2 = new EllipsoidRhumbLine(initial, rhumb1.end, ellipsoid);
 
         expect(rhumb1.heading).toEqualEpsilon(rhumb2.heading, CesiumMath.EPSILON12);
         expect(rhumb1.surfaceDistance).toEqualEpsilon(rhumb2.surfaceDistance, CesiumMath.EPSILON6);
@@ -452,7 +452,7 @@ defineSuite([
         var final = new Cartographic(CesiumMath.PI_OVER_TWO, 0.0);
         var halfway = new Cartographic(CesiumMath.PI_OVER_FOUR, 0.0);
 
-        var rhumb = new EllipsoidRhumb(initial, final, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(initial, final, ellipsoid);
         var interpolatedPoint = rhumb.interpolateUsingFraction(0.5);
 
         expect(halfway.longitude).toEqualEpsilon(interpolatedPoint.longitude, CesiumMath.EPSILON12);
@@ -467,7 +467,7 @@ defineSuite([
 
         var distance = ellipsoid.maximumRadius * CesiumMath.PI_OVER_FOUR;
 
-        var rhumb = new EllipsoidRhumb(initial, final, ellipsoid);
+        var rhumb = new EllipsoidRhumbLine(initial, final, ellipsoid);
         var interpolatedPoint = rhumb.interpolateUsingSurfaceDistance(distance);
 
         expect(halfway.longitude).toEqualEpsilon(interpolatedPoint.longitude, CesiumMath.EPSILON12);
@@ -480,7 +480,7 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var end = new Cartographic(thirtyDegrees, thirtyDegrees);
 
-        var rhumb = new EllipsoidRhumb(start, end);
+        var rhumb = new EllipsoidRhumbLine(start, end);
         var distance = rhumb.surfaceDistance;
 
         var first = rhumb.interpolateUsingSurfaceDistance(0.0);
@@ -500,7 +500,7 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var expectedMid = new Cartographic(thirtyDegrees, 0.0);
 
-        var rhumb = new EllipsoidRhumb(start, end);
+        var rhumb = new EllipsoidRhumbLine(start, end);
         var distance = Ellipsoid.WGS84.radii.x * fifteenDegrees;
 
         var midpoint = rhumb.interpolateUsingSurfaceDistance(distance);
@@ -515,7 +515,7 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var end = new Cartographic(thirtyDegrees, thirtyDegrees);
 
-        var rhumb = new EllipsoidRhumb(start, end);
+        var rhumb = new EllipsoidRhumbLine(start, end);
 
         var first = rhumb.interpolateUsingFraction(0);
         var last = rhumb.interpolateUsingFraction(1);
@@ -534,7 +534,7 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var expectedMid = new Cartographic(thirtyDegrees, 0.0);
 
-        var rhumb = new EllipsoidRhumb(start, end);
+        var rhumb = new EllipsoidRhumbLine(start, end);
 
         var midpoint = rhumb.interpolateUsingFraction(0.5);
 
@@ -550,7 +550,7 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var expectedMid = new Cartographic(thirtyDegrees, 0.0);
 
-        var rhumb = new EllipsoidRhumb(start, end);
+        var rhumb = new EllipsoidRhumbLine(start, end);
         var result = new Cartographic();
         var midpoint = rhumb.interpolateUsingFraction(0.5, result);
         expect(result).toBe(midpoint);
@@ -567,7 +567,7 @@ defineSuite([
         var thirtyDegrees = Math.PI / 6;
         var expectedMid = new Cartographic(thirtyDegrees, 0.0);
 
-        var rhumb = new EllipsoidRhumb(start, end);
+        var rhumb = new EllipsoidRhumbLine(start, end);
         var distance = Ellipsoid.WGS84.radii.x * fifteenDegrees;
 
         var result = new Cartographic();